Fix handler edges to closed generic events; share generic display naming
- Add GenericTypeDisplayNames (canonical CLR + C# display strings)
- Remap event handler Handles from open generic to closed event node when
  IEventHandler<T> provides the closed event type
- Use ResolveEmissionEventKey for Handles relationships so TargetType matches
  closed DomainEventNode.FullName
- Regression: OrganizationDeletedHandler Handles edge targets closed event
